backend/requirements.txt
flask==2.3.3
flask-mail==0.9.1
flask-cors==4.0.0
python-dotenv==1.0.0
